Refactor scoring functions to their own module
