+++ cms/server/admin/handlers/contestranking.py
@@ -3,7 +3,7 @@
 
 # Contest Management System - http://cms-dev.github.io/
 # Copyright © 2010-2013 Giovanni Mascellani <mascellani@poisson.phc.unipi.it>
-# Copyright © 2010-2015 Stefano Maggiolo <s.maggiolo@gmail.com>
+# Copyright © 2010-2018 Stefano Maggiolo <s.maggiolo@gmail.com>
 # Copyright © 2010-2012 Matteo Boscariol <boscarim@hotmail.com>
 # Copyright © 2012-2014 Luca Wehrstedt <luca.wehrstedt@gmail.com>
 # Copyright © 2014 Artem Iglikov <artem.iglikov@gmail.com>
@@ -41,7 +41,7 @@ import io
 from sqlalchemy.orm import joinedload
 
 from cms.db import Contest
-from cms.grading import task_score
+from cms.grading.scoring import task_score
 
 from .base import BaseHandler, require_permission
 
@@ -130,12 +130,12 @@ class RankingHandler(BaseHandler):
                 if show_teams:
                     row.append(p.team.name if p.team else "")
                 assert len(contest.tasks) == len(p.scores)
-                for t_score, t_partial in p.scores: # Custom field, see above
+                for t_score, t_partial in p.scores:  # Custom field, see above
                     row.append(t_score)
                     if include_partial:
                         row.append("*" if t_partial else "")
 
-                total_score, partial = p.total_score # Custom field, see above
+                total_score, partial = p.total_score  # Custom field, see above
                 row.append(total_score)
                 if include_partial:
                     row.append("*" if partial else "")
