+++ cms/server/admin/handlers/dataset.py
@@ -45,7 +45,7 @@ import tornado.web
 
 from cms.db import Dataset, Manager, Message, Participation, \
     Session, Submission, Task, Testcase
-from cms.grading import compute_changes_for_dataset
+from cms.grading.scoring import compute_changes_for_dataset
 from cmscommon.datetime import make_datetime
 from cmscommon.importers import import_testcases_from_zipfile
